app/requirements.txt
streamlit==1.36.0
elasticsearch==8.14.0
tqdm==4.66.4
openai==1.37.0
sentence_transformers==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4

--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.3.1+cpu